+++ jubatus/core/fv_converter/dynamic_binary_feature.hpp
@@ -39,7 +39,7 @@ class dynamic_binary_feature : public binary_feature {
   void add_feature(
       const std::string& key,
       const std::string& value,
-      std::vector<std::pair<std::string, float> >& ret_fv) const;
+      std::vector<std::pair<std::string, double> >& ret_fv) const;
 
  private:
   dynamic_loader loader_;
