+++ jubatus/core/fv_converter/test_binary_feature.cpp
@@ -29,7 +29,7 @@ class my_binary_feature : public core::fv_converter::binary_feature {
   void add_feature(
       const std::string& key,
       const std::string& value,
-      std::vector<std::pair<std::string, float> >& ret_fv) const {
+      std::vector<std::pair<std::string, double> >& ret_fv) const {
     ret_fv.push_back(std::make_pair(key, value.size()));
   }
 };
